fix chalk
