+++ lib/config.mjs
@@ -1,13 +1,13 @@
 import { existsSync, mkdirSync, readFileSync, statSync, writeFileSync } from 'node:fs';
 import { homedir } from 'node:os';
 import path from 'node:path';
-import chalk from 'chalk';
+import { Chalk } from 'chalk';
 
 const configDir = path.resolve(homedir(), '.config', 'fanyi');
 const configPath = path.resolve(configDir, '.fanyirc');
 
 // 初始化一个带颜色的 chalk 实例
-const chalkInstance = new chalk.Instance({ level: 3 });
+const chalk = new Chalk({ level: 3 });
 
 const config = {
   async load(path = configPath) {
@@ -29,9 +29,7 @@ const config = {
     const content = JSON.stringify(mergedOptions, null, 2);
     existsSync(configDir) || mkdirSync(configDir, { recursive: true });
     writeFileSync(path, content);
-    console.log(
-      `${chalkInstance.bgGreen(JSON.stringify(options))} config saved at ${chalkInstance.gray(path)}:`,
-    );
+    console.log(`${chalk.bgGreen(JSON.stringify(options))} config saved at ${chalk.gray(path)}:`);
     for (const [key, value] of Object.entries(options)) {
       console.log(`${chalk.cyan(key)}: ${chalk.yellow(value)}`);
     }
